clean code for Fury ArrayAsList
