+++ java/fury-core/src/main/java/org/apache/fury/serializer/collection/FuryArrayAsListSerializer.java
@@ -19,7 +19,10 @@
 
 package org.apache.fury.serializer.collection;
 
+import java.util.AbstractList;
 import java.util.Collection;
+import java.util.Iterator;
+import java.util.RandomAccess;
 import org.apache.fury.Fury;
 import org.apache.fury.annotation.Internal;
 import org.apache.fury.memory.MemoryBuffer;
@@ -28,7 +31,8 @@ import org.apache.fury.type.Type;
 /** Serializer for {@link ArrayAsList}. Helper for serialization of other classes. */
 @Internal
 @SuppressWarnings("rawtypes")
-public final class FuryArrayAsListSerializer extends CollectionSerializer<ArrayAsList> {
+public final class FuryArrayAsListSerializer
+    extends CollectionSerializer<FuryArrayAsListSerializer.ArrayAsList> {
   public FuryArrayAsListSerializer(Fury fury) {
     super(fury, ArrayAsList.class, true);
   }
@@ -42,5 +46,92 @@ public final class FuryArrayAsListSerializer extends CollectionSerializer<ArrayA
     int numElements = buffer.readVarUint32Small7();
     setNumElements(numElements);
     return new ArrayAsList(numElements);
+  }
+
+  /**
+   * A List which wrap a Java array like `java.util.Arrays.ArrayList`, but allow to replace wrapped
+   * array. Used for serialization only, do not use it in other scenarios.
+   */
+  @Internal
+  public static class ArrayAsList<E> extends AbstractList<E>
+      implements RandomAccess, java.io.Serializable {
+    private static final Object[] EMPTY = new Object[0];
+
+    private E[] array;
+    private int size;
+
+    @SuppressWarnings("unchecked")
+    public ArrayAsList(int size) {
+      array = (E[]) new Object[size];
+    }
+
+    @Override
+    public int size() {
+      return size;
+    }
+
+    @Override
+    public boolean add(E e) {
+      array[size++] = e;
+      return true;
+    }
+
+    @Override
+    public E get(int index) {
+      return array[index];
+    }
+
+    @SuppressWarnings("unchecked")
+    public void clearArray() {
+      size = 0;
+      array = (E[]) EMPTY;
+    }
+
+    public void setArray(E[] a) {
+      array = a;
+      size = a.length;
+    }
+
+    public E[] getArray() {
+      return array;
+    }
+
+    @Override
+    public E set(int index, E element) {
+      throw new UnsupportedOperationException();
+    }
+
+    @Override
+    public int indexOf(Object o) {
+      throw new UnsupportedOperationException();
+    }
+
+    @Override
+    public boolean contains(Object o) {
+      throw new UnsupportedOperationException();
+    }
+
+    /** Returns original array without copy. */
+    @Override
+    public Object[] toArray() {
+      return array;
+    }
+
+    @Override
+    public Iterator<E> iterator() {
+      return new Iterator<E>() {
+        private int index;
+
+        @Override
+        public boolean hasNext() {
+          return index < array.length;
+        }
+
+        @Override
+        public E next() {
+          return array[index++];
+        }
+      };
+    }
   }
 }
